Fix DJUI lobby list not populating: uint64_t/number ABI mismatch
web_lobby_on_room() was declared with uint64_t lobbyId but called
from JS via ccall with type 'number' (32-bit). On WASM, uint64_t
is passed as two 32-bit values, so all subsequent parameters
(connections, maxConnections, version, hostName, description) were
shifted and read as garbage. Changed to int params matching JS
number type, cast to uint64_t/uint16_t for the C callback.

+++ src/pc/web/web_lobby.c
@@ -14,10 +14,10 @@ static WebLobbyQueryCallbackPtr sQueryCallback = NULL;
 static WebLobbyQueryFinishCallbackPtr sQueryFinishCallback = NULL;
 
 EMSCRIPTEN_KEEPALIVE
-void web_lobby_on_room(uint64_t lobbyId, uint16_t connections, uint16_t maxConnections,
+void web_lobby_on_room(int lobbyId, int connections, int maxConnections,
                        const char* version, const char* hostName, const char* description) {
     if (sQueryCallback) {
-        sQueryCallback(lobbyId, 0, connections, maxConnections,
+        sQueryCallback((uint64_t)lobbyId, 0, (uint16_t)connections, (uint16_t)maxConnections,
                        "sm64coopdx", version, hostName, "", description);
     }
 }
